Refactor monitor_data backend service
